+++ changelog.md
@@ -3,6 +3,9 @@
 ## 1.25.1 (TBD)
 
 - Fixed enemy killed by [REDACTED] not being remembered when reloading a midgame save
+- Lowered volume on some of Rogo's sounds
+- Fixed charging phase of Pogo Stick bounce counting as airborne, allowing the player to hover over pits
+- Fixed collision jankiness when charging and landing Pogo Stick's super bounce immediately after taking damage
 
 ## 1.25.0 (2025-05-03)