+++ changelog.md
@@ -1,12 +1,12 @@
 # Changelog
 
-## 1.25.1 (TBD)
+## 1.25.1 (2025-05-04)
 
 - Fixed enemy killed by [REDACTED] not being remembered when reloading a midgame save
-- Lowered volume on some of Rogo's sounds
+- Fixed not being able to swing Racket Launcher at extant tennis balls when out of ammo
 - Fixed charging phase of Pogo Stick bounce counting as airborne, allowing the player to hover over pits
 - Fixed collision jankiness when charging and landing Pogo Stick's super bounce immediately after taking damage
-- Fixed not being able to swing Racket Launcher at extant tennis balls when out of ammo
+- Lowered volume on some of Rogo's sounds
 
 ## 1.25.0 (2025-05-03)